+++ scripts/discrete_representation/voxelization/volume_model_matrix_based_voxelization.py
@@ -4,7 +4,7 @@ Voxelization of a volume model using MatrixBasedVoxelization.
 In MatrixBasedVoxelization, voxels are stored as a 3D numpy array of np.bool_.
 """
 import volmdlr
-from volmdlr.core import VolumeModel
+from volmdlr.model import VolumeModel
 from volmdlr.discrete_representation import MatrixBasedVoxelization
 from volmdlr.primitives3d import Cylinder, Sphere