+++ scripts/mesh/geo_file_1.py
@@ -24,7 +24,7 @@ outer_profile = vm.wires.Contour2D.from_points(points)
 
 profile=primitives3d.ExtrudedProfile(vm.O3D, vm.Y3D, vm.Z3D, outer_profile, [], vm.X3D*0.1, name = 'extrusion')
 
-model=vm.core.VolumeModel([profile])
+model=vm.model.VolumeModel([profile])
 model.to_geo(file_name = 'model_1_geo',
              factor = 0.5,
              curvature_mesh_size = 0,