+++ scripts/stl/workflow_generic_stl_reading.py
@@ -14,7 +14,7 @@ method_volumemodel = ModelMethod(to_volumemodel_method_type, name='VolumeModel')
 
 cadview_block = CadView(name='Display3D')
 
-export_html = Export(method_type=MethodType(vm.core.VolumeModel, 'to_html_stream'), filename='VolumeModel.html',
+export_html = Export(method_type=MethodType(vm.model.VolumeModel, 'to_html_stream'), filename='VolumeModel.html',
                      extension='html', text=True, name='Export_html')
 
 pipes = [